refactoring code for users, categories and roles, dealing with network using context

+++ helpers/helper.pagination.go
@@ -7,12 +7,12 @@ import (
 )
 
 type paginationOptions struct {
-	Limit       int    `json:"limit" default:"10"`
-	Offset      int    `json:"offset" default:"0"`
+	Limit       uint   `json:"limit" default:"10"`
+	Offset      uint   `json:"offset" default:"0"`
 	Sort        string `json:"sort" default:"asc"`
-	Count       int    `json:"count"`
-	CurrentPage int    `json:"current_page" default:"1"`
-	TotalPage   int    `json:"total_page"`
+	Count       uint   `json:"count"`
+	CurrentPage uint   `json:"current_page" default:"1"`
+	TotalPage   uint   `json:"total_page"`
 }
 
 func Pagination(pagination interface{}, count int) paginationOptions {
@@ -26,11 +26,11 @@ func Pagination(pagination interface{}, count int) paginationOptions {
 	offset, _ := strconv.Atoi(payload["offset"].(string))
 	current_page, _ := strconv.Atoi(payload["current_page"].(string))
 
-	pagin.Limit = limit
-	pagin.Offset = offset
+	pagin.Limit = uint(limit)
+	pagin.Offset = uint(offset)
 	pagin.Sort = payload["sort"].(string)
-	pagin.Count = count
-	pagin.CurrentPage = current_page
+	pagin.Count = uint(count)
+	pagin.CurrentPage = uint(current_page)
 
 	if pagin.CurrentPage > 1 {
 		pagin.Offset = (pagin.Limit * pagin.CurrentPage) - pagin.Limit
@@ -38,7 +38,7 @@ func Pagination(pagination interface{}, count int) paginationOptions {
 		pagin.Offset = 0
 	}
 
-	pagin.TotalPage = int(math.Abs(math.Ceil(float64(pagin.Count) / float64(pagin.Limit))))
+	pagin.TotalPage = uint(math.Abs(math.Ceil(float64(pagin.Count) / float64(pagin.Limit))))
 
 	return pagin
 }
